pdf/requirements.txt
Django==4.2.1
djangorestframework==3.14.0
whitenoise==6.4.0
Python 3.11.3



